NUMBERS-188: refactored Complex instance methods as static functions
